requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.6.0
torch
chromadb>=0.5.0
google-genai>=0.3.0
python-dotenv>=1.0.0
rapidfuzz>=3.6.0
